+++ CHANGELOG.md
@@ -6,6 +6,10 @@ NOTE: [`epaint`](crates/epaint/CHANGELOG.md), [`eframe`](crates/eframe/CHANGELOG
 
 ## Unreleased
 
+### Added ⭐
+* Add `Spacing::slider_thickness` to globally set thickness (not width/length!) of `Slider`.
+* Add `Slider::thickness` to set thickness per `Slider`.
+* Add `Slider::circle_size` to set the circle size of a `Slider`.
 
 ## 0.21.0 - 2023-02-08 - Deadlock fix and style customizability
 * ⚠️ BREAKING: `egui::Context` now use closures for locking ([#2625](https://github.com/emilk/egui/pull/2625)):